scripts/requirements.txt
streamlit
ultralytics
torch
torchvision
torchaudio
opencv-python-headless
numpy
pillow
easyocr
joblib
pandas
nest_asyncio
scikit-learn
